server-python/pyproject.toml
[tool.poetry]
name = "server-python"
version = "0.1.0"
description = ""
authors = ["Kyle Bjordahl <kyle.bjordahl@ohmconnect.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["debug-server"], version = "^0.96.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
